Fixing customcases

+++ tests/nfs/nfs_run_spectrum_scale_upstream_file_ops_in_loop.py
@@ -14,7 +14,7 @@ def run(ceph_cluster, **kw):
     cmds = ["rm -rf ci-tests/",
             "yum install -y git wget",
             "git clone https://github.com/aravindrrh/ci-tests; cd ci-tests; git checkout scale_downstream",
-            "sh ci-tests/build_scripts/common/basic-storage-scale-custom-repo.sh",
+            # "sh ci-tests/build_scripts/common/basic-storage-scale-custom-repo.sh",
             f'echo "export SERVER=\"{server.ip_address}\"" >> ~/.bashrc && source ~/.bashrc',
             f'echo "export EXPORT=\"/ibm/scale_volume\"" >> ~/.bashrc && source ~/.bashrc',
             f'echo "export YUM_REPO=\"http://magna002.ceph.redhat.com/ceph-qe-logs/manim/repo/nfs-ganesha-v7.repo\"" >> ~/.bashrc && source ~/.bashrc']
@@ -42,27 +42,38 @@ def run(ceph_cluster, **kw):
 dir="$1"
 
 if [ -z "$dir" ]; then
- echo "Usage: $0 <directory>"
- exit 1
+  echo "Usage: $0 <directory>"
+  exit 1
 fi
 
+start_time=$(date +%s)
+duration=300 # 3600  # 1 hour
+
 while true; do
- dd if=/dev/random of="$dir/testfile.txt" bs=1k count=1
- echo "Created file testfile.txt"
+  elapsed_time=$(( $(date +%s) - start_time ))
 
- # Listing contents quietly
- ls -lrt "$dir" > /dev/null
+  if [ $elapsed_time -gt $duration ]
+  then
+    echo "Exiting ater $elapsed_time seconds"
+    sync
+    exit 0
+  fi
 
- rm -f "$dir/testfile.txt"
- echo "Deleted file testfile.txt"
+  dd if=/dev/urandom of="$dir/testfile.txt" bs=1k count=1
+  echo "Created file testfile.txt"
 
+  ls -lrt "$dir" > /dev/null
+
+  rm -f "$dir/testfile.txt"
+  echo "Deleted file testfile.txt"
 done
+exit 0
 """
     cmd = f"touch {mount_path}/cr_rm_loop.sh"
     server.exec_command(cmd=cmd, sudo=True, long_running=True)
     with server.remote_file(sudo=True, file_name=f"{mount_path}/cr_rm_loop.sh", file_mode="w") as _f:
         _f.write(test_file)
-    out = server.exec_command(cmd=f"sh {mount_path}/cr_rm_loop.sh {mount_path}", sudo=True, long_running=True)
+    out = server.exec_command(cmd=f"sh {mount_path}/cr_rm_loop.sh {mount_path} > /dev/null 2>&1; echo \"Done... Exiting script\"", sudo=True, long_running=True)
     log.info(out)
 
     return 0
